Add tests for latest Django (3.2 LTS). Add testing of Django versions >=2.2 using latest Python (3.9). Update `conftest.py` to pass tests.
To get Django 3.2 tests to work, a `SECRET_KEY` (which seems to be a requirement for Django 3.2 to run) had to be defined in `conftest.py` .  The addition of this `SECRET_KEY` also fixes the failing `py36-lint` tests, which are being tested against the latest version of Django (also currently 3.2).

Fix uses the same function (`get_random_secret_key()`) that Django uses to auto-generate `SECRET_KEY`s.